Feat: string initialized jumbo nums.
Strings overcome the limitations of C integers
in terms of representing arbitrarily big nums.
The last test case shows this, but we had to
change the capacity macro to match it. We now
have to implement dynamic allocation to change
the size of a jumbo num on a needed basis.